Add: support for Lambda@Edge
Main issue here is that those Lambda Functions need to be in
us-east-1, while the rest of our stack is not. So we need a bit
of cross-region chatter to make this happen.

One stack "LambdaEdge" is now deployed in us-east-1, with all the
functions requested. It adds the Lambda function ARN + Version to
the ParameterStore, which can be used in other Stacks to enable
the Lambda@Edge functionality (in CloudFront).