Fix: revery SecurityGroup related changes in CDK 1.21
CDK 1.21 changed how a SecurityGroup from the ECS was applied on
the ALB. Normally this was done in the ALB stack, but this is
now changed to be in the Stack that defines a EC2Service.

The way we redeploy services, this simply isn't possible (the
IAM user has no permissions to do anything with SecurityGroups).

To work around this problem, we now first remove the SecurityGroup
from the current stack, and re-add it to the ALB stack. This
basically reverts the behaviour to CDK 1.20-.