Fix: [CDN] also listen on openttd-cdn.org for CloudFront
This also required to get a certificate for that domain, which has
to be a SAN of the cdn.openttd.org. CloudFront only accepts a
single certificate per distribution.

This has one added benefit, that the connection between CloudFlare
and CloudFront can be a strict-HTTPS connection, as the certificates
validate on both ends.