Add: remove the "aws.openttd.org" domain alias
As we now run on the DNS of AWS itself, there is no longer any
need for aws.openttd.org to exist. Before it was used to CNAME
from the old DNS to AWS DNS, so:

 www.openttd.org -> www.aws.openttd.org -> ALB

As AWS services update the AWS DNS, that was needed in order to
work with deployments from CDK. As we are now migrated to Route53
as our primary DNS, nothing is CNAMEing to aws.openttd.org anymore.