Fix: remove AWSALB cookie from our traffic
Although it can be useful to have sessions to be sticky, it also
means we get questions why we have a cookie in all our requests.
As we rather prevent showing a cookiebanner telling about it, it
is just easier to remove it.

During a deployment a user can flip/flop between a new version
of the site and an old version, which might give unexpected
results. As these moments are far apary, the impact should be
very very minimal.