Fix: rollover services that can have only 1 instance running better
Before, for a short period of time, there were in fact 2 instances
running. It was possible, and it even happened, that data was
overwritten because of this.

Prevent it by making sure the old instance is shut down before a
new one is started. This will result in downtime, but it is better
than dataloss.